+++ src/stc/debug.cpp
@@ -11,11 +11,11 @@
 #include <wex/core/core.h>
 #include <wex/core/log.h>
 #include <wex/core/regex.h>
+#include <wex/factory/bind.h>
 #include <wex/factory/defs.h>
 #include <wex/factory/process.h>
 #include <wex/stc/debug.h>
 #include <wex/stc/stc.h>
-#include <wex/ui/bind.h>
 #include <wex/ui/frame.h>
 #include <wex/ui/item-dialog.h>
 #include <wex/ui/listview.h>