+++ src/ui/listview.cpp
@@ -14,11 +14,11 @@
 #include <wex/core/regex.h>
 #include <wex/core/tokenize.h>
 #include <wex/data/stc.h>
+#include <wex/factory/bind.h>
 #include <wex/factory/defs.h>
 #include <wex/factory/lexers.h>
 #include <wex/factory/printing.h>
 #include <wex/factory/stc.h>
-#include <wex/ui/bind.h>
 #include <wex/ui/frame.h>
 #include <wex/ui/frd.h>
 #include <wex/ui/item-dialog.h>