+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 **v22.10**
 - added blame revision
 - ex commands more closely follow
-    The Open Group Base Specifications Issue 7, 2018 edition
+  The Open Group Base Specifications Issue 7, 2018 edition
+- ex command-line is improved, it is now a factory::stc
 
 **v22.04** *March 6, 2022*
 - Microsoft Visual Studio 2022